🚚 refactor: refactor to the useChatStore usage
