+++ src/app/chat/features/ChatInput/ActionBar/Token/TokenTag.tsx
@@ -4,19 +4,25 @@ import { useTranslation } from 'react-i18next';
 
 import { ModelTokens } from '@/const/modelTokens';
 import { useTokenCount } from '@/hooks/useTokenCount';
+import { useChatStore } from '@/store/chat';
+import { chatSelectors } from '@/store/chat/selectors';
 import { useSessionStore } from '@/store/session';
-import { agentSelectors, chatSelectors } from '@/store/session/selectors';
+import { agentSelectors } from '@/store/session/selectors';
 import { LanguageModel } from '@/types/llm';
 
 const Token = memo(() => {
   const { t } = useTranslation('chat');
 
-  const [input, messageString, systemRole, model] = useSessionStore((s) => [
+  const [input, messageString] = useChatStore((s) => [
     s.inputMessage,
     chatSelectors.chatsMessageString(s),
+  ]);
+
+  const [systemRole, model] = useSessionStore((s) => [
     agentSelectors.currentAgentSystemRole(s),
     agentSelectors.currentAgentModel(s) as LanguageModel,
   ]);
+
   const inputTokenCount = useTokenCount(input);
 
   const systemRoleToken = useTokenCount(systemRole);
